requirements.txt
Flask==2.2.5
numpy==1.24.4
pandas==1.5.3
scikit-learn==1.2.0

gunicorn
kafka-python
joblib
apache-airflow==2.5.3